+++ gate-service/src/main/java/com/clawai/gatedemo/gate/router/HandlerRegistry.java
@@ -8,33 +8,33 @@ import java.util.concurrent.ConcurrentHashMap;
  */
 public class HandlerRegistry {
 
-    private final Map<Short, MessageHandler> handlers = new ConcurrentHashMap<>();
+    private final Map<Integer, MessageHandler> handlers = new ConcurrentHashMap<>();
 
     /**
      * @param messageId TCP 二进制协议中的消息号
      * @param handler   处理该消息号的处理器（覆盖同 ID 已有注册）
      */
-    public void register(short messageId, MessageHandler handler) {
+    public void register(int messageId, MessageHandler handler) {
         handlers.put(messageId, handler);
     }
 
     /** @param messageId 要移除的消息号 */
-    public void unregister(short messageId) {
+    public void unregister(int messageId) {
         handlers.remove(messageId);
     }
 
     /** @return 已注册的处理器，未注册时返回 null */
-    public MessageHandler getHandler(short messageId) {
+    public MessageHandler getHandler(int messageId) {
         return handlers.get(messageId);
     }
 
     /** @return 是否已注册该消息号 */
-    public boolean hasHandler(short messageId) {
+    public boolean hasHandler(int messageId) {
         return handlers.containsKey(messageId);
     }
 
     /** @return 当前注册表的浅拷贝快照，避免外部直接修改内部 Map */
-    public Map<Short, MessageHandler> getAllHandlers() {
+    public Map<Integer, MessageHandler> getAllHandlers() {
         return new ConcurrentHashMap<>(handlers);
     }
